solves issue #17
the try-except block before will never match cause subprocess dont raise it on shell return-code >= 1.
The multiple returns of the function is not an good style anyway. Better return only once in it. 
The success-message should be green, i dont find the right css file or snippet that do that.